Completed CRUD APIS

+++ src/main/java/com/nepalyatra/repositories/UserRepository.java
@@ -3,8 +3,5 @@ package com.nepalyatra.repositories;
 import com.nepalyatra.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-
-
+public interface UserRepository extends JpaRepository<User, Long> {
 }
